Create a class to store differential expression analysis results
