+++ inmoose/diffexp/__init__.py
@@ -1,0 +1,1 @@
+from .DEResults import DEResults as DEResults
