+++ setup.py
@@ -70,6 +70,7 @@ setup(
         "inmoose/sim",
         "inmoose/utils",
         "inmoose/deseq2",
+        "inmoose/diffexp",
     ],
     package_data={
         "inmoose/edgepy/edgepy_cpp": [
